+++ SnakeGame/Object.h
@@ -3,8 +3,8 @@
 
 class Object
 {
-	int m_x;
-	int m_y;
+	float m_x;
+	float m_y;
 	int m_size = 10;
 	int m_speed = 20;
 
@@ -12,17 +12,17 @@ class Object
 	bool m_isalive = true;
 public:
 	Object() {};
-	Object(int x, int y, COLORREF c);
+	Object(float x, float y, COLORREF c);
 
 	void Draw(HDC hdc);
 
-	int GetX() const { return m_x; };
-	int GetY() const { return m_y; };
+	float GetX() const { return m_x; };
+	float GetY() const { return m_y; };
 	int GetSpeed() const { return m_speed; };
 	int GetSize() const { return m_size; };
 	COLORREF GetColor() { return m_color; };
 
-	void SetPos(int x, int y) { m_x = x, m_y = y; };
+	void SetPos(float x, float y) { m_x = x, m_y = y; };
 	void SetColor(COLORREF color) { m_color = color; };
 	void SetDie() { m_isalive = false; };
